// src/types/chainData.d.ts
export interface IchainDataState {
    data: {
        name: string;
        logo: string;
        splash_screen: string;
        private_policy: string;
        terms_n_conditions: string;
        about_us: string;
        contact_us: string;
        logo_gif_dark: string;
        logo_gif_light: string;
        signup_bg: string;
        signin_bg: string;
        google_play_store: string;
        app_store: string;
    },

}